[CSHW-3] 3 ДЗ

See merge request 78945qw/Homework!4

+++ src/Ozon.Route256.Five.OrderService/Contracts/GetOrders/GetOrdersResponseItem.cs
@@ -7,9 +7,9 @@ namespace Ozon.Route256.Five.OrderService.Contracts.GetOrders;
 /// Представление заказа
 /// </summary>
 public record GetOrdersResponseItem(
-    int Id,
-    uint ItemsCount,
-    double TotalPrice,
+    long Id,
+    int ItemsCount,
+    decimal TotalPrice,
     double TotalWeight,
     string OrderType,
     DateTime OrderedAt,
